fix: add check for existence of runtime package in user project
