+++ src/index.js
@@ -4,7 +4,7 @@ const ensureNoCompetingPlugin = require('./helpers/ensureNoCompetingPlugin')
 const fixOutputDir = require('./helpers/fixOutputDir')
 const getAngularJson = require('./helpers/getAngularJson')
 const getAngularRoot = require('./helpers/getAngularRoot')
-const getAngularVersion = require('./helpers/getAngularVersion')
+const { getAngularVersion } = require('./helpers/getPackageVersion')
 const { fixServerTs, revertServerTsFix } = require('./helpers/serverModuleHelpers')
 const { getProject, setUpEdgeFunction } = require('./helpers/setUpEdgeFunction')
 const setUpHeaders = require('./helpers/setUpHeaders')
